+++ client/lua/ge/extensions/highbeam.lua
@@ -14,16 +14,40 @@ local config       -- highbeam/config.lua
 
 M.onExtensionLoaded = function()
   log('I', logTag, 'HighBeam extension loaded')
+
+  connection = require("highbeam/connection")
+  protocol   = require("highbeam/protocol")
+  vehicles   = require("highbeam/vehicles")
+  state      = require("highbeam/state")
+  chat       = require("highbeam/chat")
+  config     = require("highbeam/config")
+
+  -- Wire subsystem cross-references
+  connection.setSubsystems(vehicles, state)
+  state.setSubsystems(connection, config)
+
+  config.load()
 end
 
 M.onExtensionUnloaded = function()
   log('I', logTag, 'HighBeam extension unloaded')
+  if connection then
+    connection.disconnect()
+  end
 end
 
 M.onUpdate = function(dtReal, dtSim, dtRaw)
   -- Network tick: process incoming, send outgoing
   if connection then
     connection.tick(dtReal)
+  end
+  -- Position sending
+  if state then
+    state.tick(dtReal)
+  end
+  -- Remote vehicle interpolation
+  if vehicles then
+    vehicles.tick(dtReal)
   end
 end